Main page: link version to the download directory.
This makes the SHASUMS file discoverable.
It may even be a good idea to link to the SHASUMS directly
in the list of file links,
to encourage integrity checking on the part of the users,
but if we don't want to confuse non-techies, at the very
least we can link to the directory containing all files.