Add Basic "How To Run A Full Node" Page
Add basic page with some general information plus instructions for
Ubuntu 14.10.

Change a link on the participate page to point to the full node page.

This page is being added as English-only for now.  I figure we can
convert it into a translation template after we have instructions for
Windows and Mac OS X, and after the instrutions have actually been
tested with the released version of Bitcoin Core 0.10.0.  (These Ubuntu
instructions were tested with RC3 and the old Ubuntu packaging.)