fix error in statement about "majority of miners"
Instead of talking about a majority of miners, it is better to talk
about a majority of available hashrate. There is no way to determine the
number of miners that exist, and it is confusing to talk about a
"majority of miners" as a consequence of this and other reasons.