remove Coinfox from resources.html
Coinfox doesn't really offer any added value. It's just blogspam with rewrites of articles by myself and others. Recent example: http://www.coinfox.info/news/5057-cory-fields-health-bitcoin is a rewrite of http://coinjournal.net/mits-cory-fields-contentiousness-in-bitcoin-is-sign-of-good-health/ where the featured image is copied.
The business model of coinfox seems to be taking the top posts from the previous day on /r/bitcoin, rewriting them, and reposting them the next day.