Add BitVol to Resources/Charts
Hello.

I think this website would be an useful metric for the Charts page. It calculates the historical volatility [1] of the Bitcoin price using the Coindesk API[2] (which will change to BitcoinAverage soon && more features will be added). Volatility is a very special metric as Bitcoin is slowing moving forward towards stability that can be seen in traditional fiat currencies, and was used as an argument against Bitcoin in the past.

[1] - http://www.ivolatility.com/help/2.html
[2] - http://www.coindesk.com/api/

I would like to hear other people's input.