buy-bitcoin: Add 'Buy Bitcoin' page
This adds a 'Buy Bitcoin' page to Bitcoin.org. This page will allow us
to begin ranking and capturing traffic on search engines for the key
phrase 'buy bitcoin' which receives upwards of 100,000 unique searches a
month on Google, alone.

This strategy was employed with our Bitcoin Exchanges page, now the
fourth most-visited page on the site for the last three months. It has
received over 1,000,000 page views since it was added and is now also
the first search result on Google when the search term 'bitcoin
exchange' is used.

Unless others object, this will be merged on Tuesday, August 15th.

Closes #1573