resources: Drop Airbitz as a 'Merchant' link
This drops Airbitz from the resources page as a place to go to find
Merchants. The link was pointing to their homepage which doesn't
appear to contain any relevant information.

It was originally added to the page back in 2015
(8308d27c2a32982b67ae7d0db529376b3f26f4c0), so perhaps the page has
changed quite a bit since then.

Unless others object, this will be merged on Wednesday, May 23rd.