en: Add missing meta description to Wallet wizard
We've noted that for the 'Choose your wallet' landing page, a meta
description is missing. Meta descriptions are the text that search
engines generally display alongside web page search results; it is the
text that describes the web page. If this text is missing, a search
engine will automatically try to extract copy from the page in order to
present an accurate description of the page alongside the search result.

In the case of the choose your wallet page, Google is erroneously
extracting the wallet feature description for "2FA" and listing it as
the search engine result description.

This adds the missing meta-description, along with a correct
description, and will be merged once tests pass.