translations/en: Improving phishing copy on scams page
This adds additional text to the [Scams](https://bitcoin.org/en/scams) page,
about phishing apps and extensions, in addition to websites. This is to help
inform users of instances where nefarious actors may try to list a fake app or
extension in well-known app stores, in an attempt to steal people's
login credentials or to trick them into disclosing their wallet seed.

This is scheduled to be merged on Friday, June 25th.